backend/requirements.txt
aiosmtplib==5.1.0
alembic==1.18.1
annotated-types==0.7.0
bcrypt==4.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
pillow==11.2.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
reportlab==4.2.5
requests==2.32.3
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.37
starlette==0.50.0
stripe==12.4.0
typing-extensions==4.12.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
violation==1.3.1
slowapi==0.1.9
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
